adhoc: fix dev env user data no show
note from gpt: The error you encountered is related to accessing a property on an object that is undefined. Specifically, the error occurs in the getIntials function of the HomeCard component when trying to read the trim property of the name argument.

The error is likely caused by user.name being undefined or not being a string in some cases. To fix the warning and error, you can add a guard clause to check if user.name is valid before proceeding with the getIntials function.